+++ ntex-util/src/channel/mod.rs
@@ -2,6 +2,7 @@
 
 mod cell;
 pub mod condition;
+pub mod inplace;
 pub mod mpsc;
 pub mod oneshot;
 pub mod pool;